+++ semana14/sistema-bancario/banco.json
@@ -1,18 +1,18 @@
 [
     {
         "usuario": {
-            "nome": "Cleiton",
-            "CPF": "200123456",
-            "dataDeNascimento": "27/05/1987"
+            "nome": "Rosana Rezende",
+            "CPF": "11820701751",
+            "dataDeNascimento": "11/03/1987"
         },
         "saldo": 0,
         "extrato": []
     },
     {
         "usuario": {
-            "nome": "Rosana",
-            "CPF": "11820701751",
-            "dataDeNascimento": "11/03/1987"
+            "nome": "Cleiton Dias",
+            "CPF": "02397770504",
+            "dataDeNascimento": "27/05/1987"
         },
         "saldo": 0,
         "extrato": []
